requirements.txt
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
RPi.GPIO==0.7.1
spidev==3.6
tomli==2.0.2
ZeroSeg==0.0.2
